Add files via upload
Everything is an unmerited gift of the merciful goodness of Allah. From one’s existence, his vocation and his redemption to every ray of sun that enlivens him with its warmth and its luster and to every drop that quenches his thirst, everything is a gratuitous gift of Allah’s inexhaustible goodness.